Upgade numpy and python version (#573)
* Upgade numpy version


* Upgrade python version


* Update changelog


* Fix failing tests


---------

+++ opensearch_py_ml/series.py
@@ -1620,7 +1620,7 @@ class Series(NDFrame):
 
         >>> s = oml.DataFrame(OPENSEARCH_TEST_CLIENT, 'flights')['Carrier']
         >>> s.nunique()
-        4
+        np.int64(4)
         """
         results = super().nunique()
         return results.squeeze()
